Add files via upload
This script is executed on a Raspberry Pi which is connected to a 8-way power relay and multiple sensors. The power relay is to toggle equipment on/off i.e. Dehumidifier, co2 regulator, Exhaust Fan, inside Fans, water pump and aircon. The sensors are used to read temperature, humidity, co2 and video feedback(still in development).